fix(PPDSC-2775): update codemods
