+++ lib/codemod/src/v7.0.0/all-default.js
@@ -1,7 +1,6 @@
-const updateAudioPlayerDefaultStyles = require('./audio-player-default-styles');
-const updateAccordionDefaultStyles = require('./accordion-default-styles');
+const updateDefaultAudioPlayerPlayPauseButtonSize = require('./audio-player-play-pause-button');
 
-const codemods = [updateAudioPlayerDefaultStyles, updateAccordionDefaultStyles];
+const codemods = [updateDefaultAudioPlayerPlayPauseButtonSize];
 
 module.exports = function transformer(file, api) {
   let source = file.source;
