+++ lib/codemod/src/v7.0.0/audio-player-play-pause-button.js
@@ -1,13 +1,13 @@
 const addProp = require('../utils/add-prop');
 
-const PROP_UPDATES = [['overrides.header.stylePreset', 'accordionHeaderOld']];
+const PROP_UPDATES = [['AudioPlayerPlayPauseButton', 'size', 'large']];
 
 module.exports = function transformer(file, api) {
   const jscodeshift = api.jscodeshift;
   const root = jscodeshift(file.source);
 
-  PROP_UPDATES.forEach(([prop, value]) => {
-    addProp(root, 'Accordion', prop, value, jscodeshift);
+  PROP_UPDATES.forEach(([component, prop, value]) => {
+    addProp(root, component, prop, value, jscodeshift);
   });
 
   return root.toSource();
