+++ src/accordion/style-presets.ts
@@ -27,12 +27,6 @@ export default {
       safariOutlineStyle: '{{outlines.safariOutlineStyleDefault}}',
     },
   },
-  accordionHeaderOld: {
-    __extends: 'accordionHeader',
-    base: {
-      backgroundColor: '{{colors.interface010}}',
-    },
-  },
   accordionPanel: {
     base: {
       borderStyle: 'none none solid none',
